// src/types/linkRecord.ts
export type Status = "verified" | "unverified" | "false-positive"
export type Jenis = "Judi" | "Pornografi" | "Penipuan"

export type LinkRecord = {
  id: number
  link: string
  jenis: Jenis
  kepedean: number
  status: Status
  tanggal: string
  lastModified: string
  reasoning: string
  image: string
  flagged: boolean
}